Fix compile errors and API misuse across services and features
- Fix duplicate 'overdue' map key in status_colors (compile error)
- Make SyncQueueEntry lifecycle fields mutable (were mutated by LocalDatabaseService)
- Replace undefined QueryBuilder with SupabaseQueryBuilder/PostgrestFilterBuilder
- Fix Supabase rpc() calls: use named params: argument
- Fix supabase_flutter v2 auth types (Session/User/AuthChangeEvent, response.user)
- Fix storage_client v2 API: getPublicUrl without .data, positional expiresIn, list(path:), drop onUploadProgress
- Fix firebase_messaging v15: remove onIosActivate/onIosRefresh
- Fix printing 5.12 API: use Printing.layoutPdf instead of print(PdfFile)
- Fix excel 4.x API: CellValue/TextCellValue/IntCellValue/DoubleCellValue, encode()
- Fix csv 6.x API: Csv().encode instead of CSV().toCsv
- Remove non-const const Rx<PageInfo> constructors
- Add missing break in switch cases (session/notification/search/top_bar/report/dashboard)
- Fix num->int clamp() assignments
- Fix Dio Response.realTime -> drop unsupported timing
- Fix error_mapper null-arg + missing dart:async import
- Fix theme seedColor and app_state accent.value
- Fix isoDate splitting and asBool -> asBoolOr

Also fix several static issues found during review.

+++ lib/core/utils/pagination.dart
@@ -77,8 +77,9 @@ class PageInfo {
   final int pageSize;
   final int total;
 
-  int get totalPages =>
-      pageSize <= 0 ? 0 : (total / pageSize).ceil().clamp(1, 1 << 31);
+  int get totalPages => pageSize <= 0
+      ? 0
+      : (total / pageSize).ceil().clamp(1, 1 << 31).toInt();
 
   bool get hasNext => page < totalPages;
